feat(frontend): chain apply change and upgrade satellite
