+++ src/frontend/src/lib/services/upgrade/upgrade.init.services.ts
@@ -52,7 +52,7 @@ export const openUpgradeModal = async ({
 	});
 };
 
-const newerReleases = async ({
+export const newerReleases = async ({
 	currentVersion,
 	segments
 }: {
